+++ demo/run_bot.py
@@ -90,6 +90,8 @@ async def register_commands(client: dyscord.client.DiscordClient, message):
 
     registration = await new_command.register_to_guild(guild)
     log.info(f'Registration: {registration}')
+    registration = await new_command.register_globally()
+    log.info(f'Registration: {registration}')
 
     new_command = Command()
     new_command.generate(
@@ -111,6 +113,8 @@ async def register_commands(client: dyscord.client.DiscordClient, message):
     new_command.validate()
 
     registration = await new_command.register_to_guild(guild)
+    log.info(f'Registration: {registration}')
+    registration = await new_command.register_globally()
     log.info(f'Registration: {registration}')
 
 
@@ -180,6 +184,8 @@ async def parse_message(message: objects.Message, raw_message, client):
 
 dyscord.helper.CommandHandler.register_guild_callback('test', command_functions.test)
 dyscord.helper.CommandHandler.register_guild_callback('complex', command_functions.complex)
+dyscord.helper.CommandHandler.register_global_callback('test', command_functions.test)
+dyscord.helper.CommandHandler.register_global_callback('complex', command_functions.complex)
 
 client._register_raw_callback(handle_any)