+++ tests/fixtures/__init__.py
@@ -1,0 +1,6 @@
+from .fixtures import mock_api, mock_websocket
+
+__all__ = [
+    'mock_api',
+    'mock_websocket'
+]